@@ -7,8 +7,8 @@
   "dependencies": {
     "@appbaseio/reactivesearch": "^2.12.1",
     "moment": "^2.20.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
