@@ -28,52 +28,47 @@ import CancelSvg from '../shared/CancelSvg';
 class DateRange extends Component {
 	constructor(props) {
 		super(props);
+
+		let currentDate = props.defaultValue || props.value || null;
+		if (props.selectedValue) {
+			if (Array.isArray(props.selectedValue)) {
+				currentDate = {
+					start: new Date(props.selectedValue[0]),
+					end: new Date(props.selectedValue[1]),
+				};
+			} else {
+				const { start, end } = props.selectedValue;
+				currentDate = {
+					start: new Date(start),
+					end: new Date(end),
+				};
+			}
+		}
+
 		this.state = {
-			currentDate: null,
+			currentDate,
 			dateHovered: null,
 		};
 		this.locked = false;
+		const hasMounted = false;
+
+		props.addComponent(props.componentId);
 		props.setQueryListener(props.componentId, props.onQueryChange, null);
-	}
+		this.setReact(props);
 
-	componentWillMount() {
-		this.props.addComponent(this.props.componentId);
-		this.setReact(this.props);
-
-		if (this.props.selectedValue) {
-			// parsing string values from selectedValue to date objects for DayPicker
-			// for value as an array from URL
-			if (Array.isArray(this.props.selectedValue)) {
-				this.handleDateChange(
-					{
-						start: new Date(this.props.selectedValue[0]),
-						end: new Date(this.props.selectedValue[1]),
-					},
-					false,
-				);
-			} else {
-				// for value as an object for SSR
-				const { start, end } = this.props.selectedValue;
-				this.handleDateChange(
-					{
-						start: new Date(start),
-						end: new Date(end),
-					},
-					false,
-				);
-			}
-		} else if (this.props.defaultSelected) {
-			this.handleDateChange(this.props.defaultSelected, false);
+		if (currentDate) {
+			this.handleDateChange(currentDate, false, props, hasMounted);
 		}
 	}
 
-	componentWillReceiveProps(nextProps) {
-		checkPropChange(this.props.react, nextProps.react, () => this.setReact(nextProps));
-		if (!isEqual(this.props.defaultSelected, nextProps.defaultSelected)) {
-			this.handleDateChange(nextProps.defaultSelected, false, nextProps);
+	componentDidUpdate(prevProps) {
+		checkPropChange(this.props.react, prevProps.react, () => this.setReact(this.props));
+
+		if (!isEqual(this.props.value, prevProps.value)) {
+			this.handleDateChange(this.props.value, false, this.props);
 		} else {
 			const { currentDate } = this.state;
-			const { selectedValue } = nextProps;
+			const { selectedValue } = this.props;
 			// comparing array format of selectedValue with object form of the state if not null
 			if (
 				!isEqual(
@@ -85,21 +80,22 @@ class DateRange extends Component {
 						: null,
 					selectedValue,
 				)
-				&& !isEqual(this.props.selectedValue, selectedValue)
+				&& !isEqual(prevProps.selectedValue, selectedValue)
 			) {
 				this.handleDateChange(
 					selectedValue
 						? {
-							start: nextProps.selectedValue[0],
-							end: nextProps.selectedValue[1],
+							start: this.props.selectedValue[0] || '',
+							end: this.props.selectedValue[1] || '',
 						} // prettier-ignore
 						: null,
 					true,
-					nextProps,
+					this.props,
 				);
 			}
 		}
-		checkPropChange(this.props.dataField, nextProps.dataField, () =>
+
+		checkPropChange(this.props.dataField, prevProps.dataField, () =>
 			this.updateQuery(
 				this.state.currentDate
 					? {
@@ -108,7 +104,7 @@ class DateRange extends Component {
 						end: formatDate(this.state.currentDate.end),
 					} // prettier-ignore
 					: this.state.currentDate,
-				nextProps,
+				this.props,
 			));
 	}
 
@@ -180,23 +176,44 @@ class DateRange extends Component {
 
 	clearDayPickerStart = () => {
 		if (this.state.currentDate && this.state.currentDate.start !== '') {
-			this.handleStartDate('', false); // resets the day picker component
+			const { value, onChange } = this.props;
+			if (value) {
+				if (onChange) onChange({ start: '', end: this.state.currentDate.end });
+			} else {
+				this.handleStartDate('', false); // resets the day picker component
+			}
 		}
 	};
 
 	clearDayPickerEnd = () => {
 		if (this.state.currentDate && this.state.currentDate.end !== '') {
 			this.handleEndDate(''); // resets the day picker component
+			const { value, onChange } = this.props;
+			if (value) {
+				if (onChange) onChange({ start: this.state.currentDate.start, end: '' });
+			} else {
+				this.handleStartDate('', false); // resets the day picker component
+			}
 		}
 	};
 
 	handleStartDate = (date, autoFocus = true) => {
 		const { currentDate } = this.state;
 		const end = currentDate ? currentDate.end : '';
-		this.handleDateChange({
-			start: date,
-			end,
-		});
+		const { value, onChange } = this.props;
+		if (value) {
+			if (onChange) {
+				onChange({
+					start: date,
+					end,
+				});
+			}
+		} else {
+			this.handleDateChange({
+				start: date,
+				end,
+			});
+		}
 		// focus the end date DayPicker if its empty
 		if (this.props.autoFocusEnd && autoFocus) {
 			// TODO: replace with a single date component in v2.1.0
@@ -206,10 +223,20 @@ class DateRange extends Component {
 
 	handleEndDate = (date) => {
 		const { currentDate } = this.state;
-		this.handleDateChange({
-			start: currentDate ? currentDate.start : '',
-			end: date,
-		});
+		const { value, onChange } = this.props;
+		if (value) {
+			if (onChange) {
+				onChange({
+					start: currentDate ? currentDate.start : '',
+					end: date,
+				});
+			}
+		} else {
+			this.handleDateChange({
+				start: currentDate ? currentDate.start : '',
+				end: date,
+			});
+		}
 	};
 
 	handleDayMouseEnter = (day) => {
@@ -218,7 +245,8 @@ class DateRange extends Component {
 		});
 	};
 
-	handleDateChange = (currentDate, isDefaultValue = false, props = this.props) => {
+	handleDateChange = (
+		currentDate, isDefaultValue = false, props = this.props, hasMounted = true) => {
 		// ignore state updates when component is locked
 		if (props.beforeValueChange && this.locked) {
 			return;
@@ -236,16 +264,22 @@ class DateRange extends Component {
 		}
 
 		const performUpdate = () => {
-			this.setState(
-				{
-					currentDate,
-				},
-				() => {
-					this.updateQuery(value, props);
-					this.locked = false;
-					if (props.onValueChange) props.onValueChange(value);
-				},
-			);
+			const handleUpdates = () => {
+				this.updateQuery(value, props);
+				this.locked = false;
+				if (props.onValueChange) props.onValueChange(value);
+			};
+
+			if (hasMounted) {
+				this.setState(
+					{
+						currentDate,
+					},
+					handleUpdates,
+				);
+			} else {
+				handleUpdates();
+			}
 		};
 		checkValueChange(props.componentId, value, props.beforeValueChange, performUpdate);
 	};
@@ -270,7 +304,7 @@ class DateRange extends Component {
 		const { currentDate, dateHovered } = this.state;
 		const start = currentDate ? currentDate.start : '';
 		const end = currentDate ? currentDate.end : '';
-		const endDay = dateHovered || '';
+		const endDay = currentDate ? dateHovered : '';
 		const selectedDays = [start, { from: start, to: endDay }];
 		const modifiers = { start, end: endDay };
 		return (
@@ -406,13 +440,16 @@ DateRange.propTypes = {
 	componentId: types.stringRequired,
 	dataField: types.dataFieldArray,
 	dayPickerInputProps: types.props,
-	defaultSelected: types.dateObject,
+	defaultValue: types.dateObject,
+	value: types.dateObject,
 	filterLabel: types.string,
 	focused: types.bool,
 	initialMonth: types.dateObject,
 	innerClass: types.style,
 	numberOfMonths: types.number,
 	onQueryChange: types.func,
+	onValueChange: types.func,
+	onChange: types.func,
 	placeholder: types.rangeLabels,
 	queryFormat: types.queryFormatDate,
 	react: types.react,
